requirements.txt
requests
requests_oauthlib
Flask==1.1.2
google-cloud-datastore==1.12.0
google-api-python-client==1.8.2
